Machine: add imx8mq-pe100a2g machine for PE100A 2G sku

+++ recipes-core/systemd/systemd-serialgetty.bbappend
@@ -1,13 +1,7 @@
 FILESEXTRAPATHS_prepend := "${THISDIR}/${BPN}:"
 
-do_install_append_imx8mq-pe100a() {
-  rm ${D}${sysconfdir}/systemd/system/getty.target.wants/serial-getty@ttymxc0.service
-}
-
-do_install_append_imx8mq-pv100a() {
-  rm ${D}${sysconfdir}/systemd/system/getty.target.wants/serial-getty@ttymxc0.service
-}
-
-do_install_append_imx8mq-pv100a2g() {
-  rm ${D}${sysconfdir}/systemd/system/getty.target.wants/serial-getty@ttymxc0.service
+do_install_append() {
+  if [ "${SERIAL_CONSOLES_ENABLE}" = "false" ]; then
+    rm ${D}${sysconfdir}/systemd/system/getty.target.wants/serial-getty@ttymxc0.service
+  fi
 }
